+++ index.js
@@ -1,109 +1,9 @@
-const defaultConfig = require('./lib/default-config')
-const generateBody = require('./lib/generate-body')
-const commitIsInPR = require('./lib/commit-is-in-pr')
-const organizeCommits = require('./lib/organize-commits')
-const generateLabel = require('./lib/generate-label')
-const reopenClosed = require('./lib/reopen-closed')
-const metadata = require('./lib/metadata')
-const {stripAt} = require('./lib/helpers')
+const openIssues = require('./lib/open-issues')
+const mergeHandler = require('./lib/merge-handler')
 
 module.exports = (robot) => {
-  robot.on('push', async context => {
-    if (!context.payload.head_commit) return
-
-    const config = await context.config('config.yml')
-    const cfg = config && config.todo ? {...defaultConfig, ...config.todo} : defaultConfig
-
-    const [labels, prs, tree] = await Promise.all([
-      generateLabel(context, cfg),
-      context.github.pullRequests.getAll(context.repo()),
-      context.github.gitdata.getTree(context.repo({sha: context.payload.head_commit.id, recursive: true}))
-    ])
-
-    if (tree.truncated) {
-      robot.log.error(new Error('Tree was too large for one recursive request.'))
-      return
-    }
-
-    const pr = commitIsInPR(context, prs)
-
-    // Get array of issue objects in the current repo
-    const {pusher, commits} = context.payload
-    const author = pusher.name
-
-    // Get the most up-to-date contents of each file
-    // by the commit it was most recently edited in.
-    const commitsByFiles = await organizeCommits(context, commits, tree)
-    commitsByFiles.forEach(async (files, commitSha) => {
-      files.forEach(async ({ contents, sha }, file) => {
-        // Get issue titles
-        const regexFlags = cfg.caseSensitive ? 'g' : 'gi'
-        const re = new RegExp(`${cfg.keyword}\\s(.*)`, regexFlags)
-        const matches = contents.match(re)
-        if (!matches) return
-
-        const titles = matches.map(title => title.replace(new RegExp(`${cfg.keyword} `, regexFlags), ''))
-
-        titles.forEach(async title => {
-          // Check if an issue with that title exists
-          const search = await context.github.search.issues({
-            q: `${title} in:title repo:${context.payload.repository.full_name}`,
-            per_page: 100
-          })
-
-          if (search.data.total_count !== 0) {
-            const existingIssue = search.data.items.find(issue => {
-              if (!issue.body) return false
-              const titleKey = metadata(context, issue).get('title')
-              const fileKey = metadata(context, issue).get('file')
-              return titleKey === title && fileKey === file
-            })
-
-            if (existingIssue) {
-              if (cfg.reopenClosed && existingIssue.state === 'closed') {
-                return reopenClosed(robot.log, context, cfg, existingIssue.number, file, commitSha)
-              } else {
-                return
-              }
-            }
-          }
-
-          const body = generateBody(context, cfg, title, file, contents, author, commitSha, pr)
-
-          if (pr) {
-            return context.github.pullRequests.createComment(context.repo({
-              number: pr,
-              body
-            }))
-          }
-
-          const issueObj = { title, body, labels }
-          if (cfg.autoAssign === true) {
-            issueObj.assignee = author
-          } else if (typeof cfg.autoAssign === 'string') {
-            issueObj.assignee = cfg.autoAssign
-          } else if (Array.isArray(cfg.autoAssign)) {
-            issueObj.assignees = cfg.autoAssign.map(n => stripAt(n))
-          }
-
-          const issue = context.issue(issueObj)
-
-          robot.log.info(`Creating issue: ${issue.title}, in ${context.repo().owner}/${context.repo().repo}`)
-          return context.github.issues.create(issue)
-        })
-      })
-    })
-  })
-
-  // :TODO: Verify PR merge event name
-  robot.on('pull_request.merged', async context => {
-    // :TODO: Create new issues from PR comments if they still exist
-    // @body This should be done after the Gitdata API has been settled, since that changes how files are parsed for todo titles.
-    // 1. Get all PR comments
-    // 2. Filter for TODO comment
-    // 3. For each TODO comment, see if TODO is still in code merged.
-    // 4. If yes, open an issue.
-  })
+  robot.on('push', async context => openIssues(context, robot))
+  robot.on('pull_request.merged', async context => mergeHandler(context, robot))
 
   robot.on('installation.created', context => {
     const repos = context.payload.repositories.reduce((prev, repo, i, arr) => {
