+++ lib/merge-handler.js
@@ -1,0 +1,7 @@
+module.exports = async (context, robot) => {
+  // 0. Get merged PR data
+  // 1. Get all comments in the merged PR
+  // 2. Filter for TODO comment (by author?)
+  // 3. For each TODO comment, see if TODO is still in code merged.
+  // 4. If yes, open an issue.
+}
